refactor: split server stats settings routes
